src/Chapter15/FIlesIO/InputSream/Account.java
package Chapter15.FIlesIO.InputSream;

import java.io.Serializable;
import java.math.BigDecimal;

public class Account implements Serializable {
    private String name;
    private BigDecimal balance;


    public Account(String name, BigDecimal balance) {
        this.name = name;
        this.balance = balance;
    }

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public BigDecimal getBalance() {
        return balance;
    }

    public void setBalance(BigDecimal balance) {
        this.balance = balance;
    }

    @Override
    public String toString() {
        return "Account{" +
                "name='" + name + '\'' +
                ", balance=" + balance +
                '}';
    }
}
